chore(cms): move location of blocks folder
