+++ src/blocks/Banner/Component.tsx
