+++ src/blocks/Banner/config.ts
