+++ src/blocks/Cover/Component.tsx
