+++ src/blocks/Cover/config.ts
