+++ src/blocks/MediaBlock/config.ts
