+++ tsconfig.json
@@ -22,14 +22,13 @@
       "@/*": ["src/*"],
       "@/access/*": ["src/access/*"],
       "@/app/*": ["src/app/*"],
-      "@/blocks/*": ["src/app/blocks/*"],
+      "@/blocks/*": ["src/blocks/*"],
       "@/collections/*": ["src/collections/*"],
-      "@/components/*": ["src/app/components/*"],
-      "@/data/*": ["src/app/data/*"],
+      "@/components/*": ["src/components/*"],
+      "@/data/*": ["src/data/*"],
       "@/images/*": ["./public/images/*"],
       "@/fields/*": ["src/fields/*"],
       "@/globals/*": ["src/globals/*"],
-      "@/lib/*": ["src/app/lib/*"],
       "@/public/*": ["./public/*"],
       "@/cssVariables": ["./cssvariables.js"],
       "@/styles/*": ["src/styles/*"],
